Do the refactor I promised: now the main svg hierarchy looks like
  svg -> g.country -> g.bar -> rect, text.barSize

The code is enjoyably clean, as well. To make my life easier, the percentile
objects now contain their country string (makes it easy to omit text for the
'percentBar-...' type 'country').

I spent a lot of time struggling to make the new data join work, and uncovered a
critical gap in my knowledge of d3. Specifically (and this is pretty obvious in
hindsight), determining which data is new and which is old is not an automagic
process. You need to tell d3 how to compare objects with the optional key
argument to .data(). In my case, the really important parts of the percentile
object are the size and positioning aspects, so that's what I keyed on.

In the process of making this change, I nearly submitted broken code. It's
become very clear that I need to get my work under test. I think the correct
plan of action is adding the slider and then shifting focus to tests.